requirement.txt
autopep8==1.6.0
cmake==3.22.2
dlib==19.23.0
face-recognition-models==0.3.0
mysql-connector==2.2.9
numpy==1.22.1
opencv-contrib-python==4.5.5.62
pandas==1.4.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
toml==0.10.2
